third_party/WebKit/LayoutTests/http/tests/media/media-source/webkitmediasource-removesourcebuffer-texttrack-expected.txt

CONSOLE WARNING: 'WebKitMediaSource' is deprecated. Please use 'MediaSource' instead.
EVENT(loadedmetadata)
EVENT(error)
END OF TEST
